♻️ Use a single route for updating name
