+++ src/pages/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
-import { navigate } from 'gatsby'
 import shortUUID from 'short-uuid'
 import { SEO } from '../components/seo'
+import { navigateToRoom } from '../utils/room'
 
 const CreateARoom = () => {
   React.useEffect(() => {
-    navigate(`/room/${shortUUID.generate()}`)
+    navigateToRoom(shortUUID.generate())
   }, [])
 
   return <SEO title="Join a room" />
