+++ src/pages/room.tsx
@@ -2,9 +2,8 @@ import React, { useState } from 'react'
 import { Router } from '@reach/router'
 import { Room } from '../room-pages/room'
 import { signIn } from '../utils/firebase'
-import { Loading } from '../room-pages/loading'
 import { Results } from '../room-pages/results'
-import { ProvideName } from '../room-pages/provide-name'
+import NotFound from './404'
 
 const RoomRouter = () => {
   const [uid, setUid] = useState('')
@@ -17,10 +16,9 @@ const RoomRouter = () => {
 
   return (
     <Router basepath="/room">
-      <ProvideName path="/:roomId/name" />
       <Results path="/:roomId/results" />
       <Room uid={uid} path="/:roomId" />
-      <Loading default />
+      <NotFound default />
     </Router>
   )
 }
