+++ src/room-pages/results.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import { navigate } from 'gatsby'
 import { PageLayout } from '../layouts/page'
 import { SEO } from '../components/seo'
 import { toRoomName } from '../utils/room-name'
 import { useActiveTicket } from '../utils/room'
 import { RouteComponentProps } from '@reach/router'
+import { navigateToRoom } from '../utils/room'
 
 type ResultsProps = RouteComponentProps<{
   roomId: string
@@ -41,9 +41,6 @@ export const Results: React.FunctionComponent<ResultsProps> = ({
   results,
 }) => {
   const [activeTicket] = useActiveTicket(roomId)
-  const startVote = () => {
-    navigate(`/room/${roomId}/vote`)
-  }
 
   const votes = Object.values(results)
   const points = votes
@@ -124,14 +121,14 @@ export const Results: React.FunctionComponent<ResultsProps> = ({
 
       <button
         className="button is-primary"
-        onClick={() => navigate(`/room/${roomId}`)}
+        onClick={() => navigateToRoom(roomId)}
       >
         Go back to home
       </button>
 
       <button
         className="button is-secondary"
-        onClick={() => navigate(`/room/${roomId}/vote`)}
+        onClick={() => navigateToRoom(roomId, 'vote')}
       >
         Start another vote round
       </button>
